Issue #3341 - Use HttpClientProvider in WebSocketCoreClient
