+++ jetty-websocket/websocket-core/src/main/java/module-info.java
@@ -37,6 +37,7 @@ module org.eclipse.jetty.websocket.core
     requires org.eclipse.jetty.io;
     requires org.eclipse.jetty.http;
     requires org.eclipse.jetty.server;
+    requires static org.eclipse.jetty.xml;
     requires org.eclipse.jetty.util;
 
     uses Extension;
